requirements.txt
Django>=5.0
Pillow
psycopg2-binary
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-filter>=23.5
drf-spectacular>=0.27.0
python-decouple>=3.8
twilio>=8.0.0
gunicorn
whitenoise[brotli]
dj-database-url
